layers/otel-common/requirements.txt
# OpenTelemetry SDK
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-semantic-conventions>=0.45b0

# OTel Exporters
opentelemetry-exporter-otlp-proto-http>=1.24.0

# AWS integrations
opentelemetry-instrumentation-aws-lambda>=0.45b0
opentelemetry-instrumentation-botocore>=0.45b0
opentelemetry-instrumentation-requests>=0.45b0

# Pydantic for attribute validation
pydantic>=2.0.0
